+++ PKHeX.Core/Editing/Bulk/ModifyResult.cs
@@ -1,0 +1,11 @@
+﻿namespace PKHeX.Core
+{
+    internal enum ModifyResult
+    {
+        None,
+        Invalid,
+        Error,
+        Filtered,
+        Modified,
+    }
+}
