+++ PKHeX.Core/Util/ReflectUtil.cs
@@ -8,53 +8,44 @@ namespace PKHeX.Core
 {
     public static class ReflectUtil
     {
-        public static bool IsValueEqual(object obj, string propertyName, object value)
+        public static bool IsValueEqual(this PropertyInfo pi, object obj, object value)
         {
-            PropertyInfo pi = obj.GetType().GetTypeInfo().GetDeclaredProperty(propertyName);
-            if (pi == null)
-                return false;
             var v = pi.GetValue(obj, null);
             var c = ConvertValue(value, pi.PropertyType);
             return v.Equals(c);
         }
-        public static void SetValue(object obj, string propertyName, object value)
+        public static void SetValue(PropertyInfo pi, object obj, object value)
         {
-            PropertyInfo pi = obj.GetType().GetTypeInfo().GetDeclaredProperty(propertyName);
-            pi.SetValue(obj, ConvertValue(value, pi.PropertyType), null);
+            var c = ConvertValue(value, pi.PropertyType);
+            pi.SetValue(obj, c, null);
         }
 
-        public static object GetValue(object obj, string propertyName)
-        {
-            PropertyInfo pi = obj.GetType().GetTypeInfo().GetDeclaredProperty(propertyName);
-            return pi.GetValue(obj, null);
-        }
-
+        public static object GetValue(object obj, string name) => GetPropertyInfo(obj?.GetType().GetTypeInfo(), name)?.GetValue(obj);
+        public static void SetValue(object obj, string name, object value) => GetPropertyInfo(obj?.GetType().GetTypeInfo(), name)?.SetValue(obj, value, null);
         public static object GetValue(Type t, string propertyName) => t.GetTypeInfo().GetDeclaredProperty(propertyName).GetValue(null);
         public static void SetValue(Type t, string propertyName, object value) => t.GetTypeInfo().GetDeclaredProperty(propertyName).SetValue(null, value);
 
         public static IEnumerable<string> GetPropertiesStartWithPrefix(Type type, string prefix)
         {
-            return type.GetTypeInfo().DeclaredProperties
+            return type.GetTypeInfo().GetAllTypeInfo().SelectMany(GetAllProperties)
                 .Where(p => p.Name.StartsWith(prefix, StringComparison.Ordinal))
                 .Select(p => p.Name);
         }
         public static IEnumerable<string> GetPropertiesCanWritePublic(Type type)
         {
-            return type.GetTypeInfo().DeclaredProperties
-                .Where(p => p.CanWrite && p.SetMethod.IsPublic)
-                .Select(p => p.Name);
+            return GetAllPropertyInfoCanWritePublic(type).Select(p => p.Name);
         }
+        public static IEnumerable<PropertyInfo> GetAllPropertyInfoCanWritePublic(Type type)
+        {
+            return type.GetTypeInfo().GetAllTypeInfo().SelectMany(GetAllProperties)
+                            .Where(p => p.CanWrite && p.SetMethod.IsPublic);
+        }
+
         public static IEnumerable<string> GetPropertiesCanWritePublicDeclared(Type type)
         {
-            return GetPropertiesCanWritePublic(type);
-        }
-        public static bool HasProperty(this Type type, string name)
-        {
-            return type.GetTypeInfo().GetDeclaredProperty(name) != null;
-        }
-        public static bool HasPropertyAll(this Type type, Type Base, string name)
-        {
-            return HasProperty(type, name) || HasProperty(Base, name);
+            return type.GetTypeInfo().GetAllProperties()
+                .Where(p => p.CanWrite && p.SetMethod.IsPublic)
+                .Select(p => p.Name);
         }
 
         private static object ConvertValue(object value, Type type)
@@ -69,9 +60,44 @@ namespace PKHeX.Core
             // Convert.ChangeType is suitable for most things
             return Convert.ChangeType(value, type);
         }
-        public static bool? GetBooleanState(object obj, string prop)
+
+        public static IEnumerable<ConstructorInfo> GetAllConstructors(this TypeInfo typeInfo)
+            => GetAll(typeInfo, ti => ti.DeclaredConstructors);
+
+        public static IEnumerable<EventInfo> GetAllEvents(this TypeInfo typeInfo)
+            => GetAll(typeInfo, ti => ti.DeclaredEvents);
+
+        public static IEnumerable<FieldInfo> GetAllFields(this TypeInfo typeInfo)
+            => GetAll(typeInfo, ti => ti.DeclaredFields);
+
+        public static IEnumerable<MemberInfo> GetAllMembers(this TypeInfo typeInfo)
+            => GetAll(typeInfo, ti => ti.DeclaredMembers);
+
+        public static IEnumerable<MethodInfo> GetAllMethods(this TypeInfo typeInfo)
+            => GetAll(typeInfo, ti => ti.DeclaredMethods);
+
+        public static IEnumerable<TypeInfo> GetAllNestedTypes(this TypeInfo typeInfo)
+            => GetAll(typeInfo, ti => ti.DeclaredNestedTypes);
+
+        public static IEnumerable<PropertyInfo> GetAllProperties(this TypeInfo typeInfo)
+            => GetAll(typeInfo, ti => ti.DeclaredProperties);
+
+        public static IEnumerable<TypeInfo> GetAllTypeInfo(this TypeInfo typeInfo)
         {
-            return obj.GetType().HasProperty(prop) ? GetValue(obj, prop) as bool? : null;
+            while (typeInfo != null)
+            {
+                yield return typeInfo;
+                typeInfo = typeInfo.BaseType?.GetTypeInfo();
+            }
+        }
+        public static bool HasProperty(object obj, string name, out PropertyInfo pi) => (pi = GetPropertyInfo(obj?.GetType().GetTypeInfo(), name)) != null;
+        public static PropertyInfo GetPropertyInfo(this TypeInfo typeInfo, string name)
+        {
+            return typeInfo.GetAllTypeInfo().Select(t => t.GetDeclaredProperty(name)).FirstOrDefault(pi => pi != null);
+        }
+        private static IEnumerable<T> GetAll<T>(this TypeInfo typeInfo, Func<TypeInfo, IEnumerable<T>> accessor)
+        {
+            return GetAllTypeInfo(typeInfo).SelectMany(y => accessor(typeInfo));
         }
     }
 }
