+++ PKHeX.WinForms/Subforms/SAV_MysteryGiftDB.cs
@@ -281,29 +281,9 @@ namespace PKHeX.WinForms
 
             if (RTB_Instructions.Lines.Any(line => line.Length > 0))
             {
-                var raw = RTB_Instructions.Lines.Where(line => !string.IsNullOrWhiteSpace(line) && new[] { '!', '=' }.Contains(line[0]));
-
-                var filters = (from line in raw
-                        let eval = line[0] == '='
-                        let split = line.Substring(1).Split('=')
-                        where split.Length == 2 && !string.IsNullOrWhiteSpace(split[0])
-                        select new BatchEditor.StringInstruction { PropertyName = split[0], PropertyValue = split[1], Evaluator = eval }).ToArray();
-
-                if (filters.Any(z => string.IsNullOrWhiteSpace(z.PropertyValue)))
-                { WinFormsUtil.Error(MsgBEFilterEmpty); return; }
-
-                res = res.Where(gift => // Compare across all filters
-                {
-                    foreach (var cmd in filters)
-                    {
-                        if (!gift.GetType().HasPropertyAll(cmd.PropertyName))
-                            return false;
-                        try { if (gift.GetType().IsValueEqual(gift, cmd.PropertyName, cmd.PropertyValue) == cmd.Evaluator) continue; }
-                        catch { Debug.WriteLine($"Unable to compare {cmd.PropertyName} to {cmd.PropertyValue}."); }
-                        return false;
-                    }
-                    return true;
-                });
+                var filters = StringInstruction.GetFilters(RTB_Instructions.Lines).ToArray();
+                BatchEditing.ScreenStrings(filters);
+                res = res.Where(pkm => IsPKMFiltered(pkm, filters)); // Compare across all filters
             }
 
             var results = res.ToArray();
@@ -313,6 +293,20 @@ namespace PKHeX.WinForms
             SetResults(new List<MysteryGift>(results)); // updates Count Label as well.
             System.Media.SystemSounds.Asterisk.Play();
         }
+
+        private static bool IsPKMFiltered(MysteryGift gift, StringInstruction[] filters)
+        {
+            foreach (var cmd in filters)
+            {
+                if (!ReflectUtil.HasProperty(gift, cmd.PropertyName, out var pi))
+                    return false;
+                try { if (pi.IsValueEqual(gift, cmd.PropertyValue) == cmd.Evaluator) continue; }
+                catch { Debug.WriteLine($"Unable to compare {cmd.PropertyName} to {cmd.PropertyValue}."); }
+                return false;
+            }
+            return true;
+        }
+
         private void UpdateScroll(object sender, ScrollEventArgs e)
         {
             if (e.OldValue != e.NewValue)
